+++ CHANGELOG.md
@@ -2893,3 +2893,4 @@
 - [2026-07-16 12:25] style(storage): harmonize light and dark theme palette for storage
 - [2026-07-16 12:51] fix(test): prevent double tap duplicate action in test
 - [2026-07-16 13:21] chore(ui): update dependencies and cleanup unused imports in ui
+- [2026-07-16 14:15] style(test): harmonize light and dark theme palette for test